Revert to working 16-bit 48kHz USB Audio configuration

+++ Inc/usbd_audio_if.h
@@ -12,7 +12,7 @@ extern USBD_AUDIO_ItfTypeDef USBD_AUDIO_fops;
 /* Audio processing buffer and counter variables */
 extern volatile uint32_t usb_audio_rx_count;
 extern volatile uint8_t usb_audio_streaming;
-extern volatile int32_t usb_audio_buffer[];
+extern volatile int16_t usb_audio_buffer[];
 
 #ifdef __cplusplus
 }
